pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.1"
description = "A URL shortener built with FastAPI"
authors = ["Anand Krishna <anandkrishna2312@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
databases = "^0.4.3"
python-dotenv = "^0.17.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
starlette-discord = "^0.1.0"
oauthlib = "^3.1.0"
aiohttp = "^3.7.4"
asyncpg = "^0.22.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pre-commit = "^2.12.0"
aiosqlite = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
